CCDM-1-Fix-Initial-Commit-Bugs: Rename files and fix directory pathways
